Treat any mounts below /media as non-essential
Do not consider mounts below /media as essential as the FHS defines this
mount point for removeable media and we shouldn't fail to boot if such
removeable media is not attached.
This can be overridden by using the "fail" mount option.

Bug-Debian: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=785525