build-sys: suppress warnings of unused results
Unfortunately, gcc keeps warning about those even when we use an
explicit (void) cast to indicate we are not interested in the result.

LLVM's clang does not have that issue and works fine with the casts.

GCC bug being tracked at:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66425

Until that GCC bug is fixed (and the version with the fix is in
many/most distributions) or we switch to LLVM as the default compiler,
it looks like we'll have to disable this warning by default...

Tested by building files known to present warnings about unused results
without the suppression, confirmed that the warnings were no longer
present with this patch applied.

This partially reverts commit 00c11bc53a1d37 ("build-sys: don't suppress irrelevant warnings").