nspawn: check if kernel supports userns as early as possible
If the kernel do not support user namespace then one of the children
created by nspawn parent will fail at clone(CLONE_NEWUSER) with a generic
error EINVAL and without logging the error. At the same time the parent
may also try to setup the user namespace and will fail with another
error.

To improve this, check if the kernel supports user namespace as early as
possible and while we are it rename the function determine_names() to
determine_container_environment(). This way we combine the checks into
the same function.