sd-netlink: respect attribute type flags
Though currently unused by us, netlink attribute types support embedding flags to indicate
if the type is encoded in network byte-order and if it is a nested attribute. Read out
these flags and handle them accordingly.

We will now refuse to enter a container unless it is marked as nested, and we will swap
the byteorder in case it is non-native when reading out integers (though this is not
needed by any of the types we currently support).