lldp: add public function to export LLDP TLV packets
Add a public function to get a list of current LLDP neighbours' TLV
packets. The function populates an array of pointers to the opaque
type sd_lldp_tlv_packet and returns the number of elements
found. Callers must take care of freeing the array and decreasing the
refcount of elements when done.