lldp: add public function to export LLDP TLVs
Add a public function to get a list of current LLDP neighbours'
TLVs. The function populates an array of pointers to the opaque type
sd_lldp_tlv and returns the number of elements found. Callers must
take care of freeing the array and decreasing the refcount of
elements when done.