nspawn: Open-code changing selinux exec context
libselinux caches state that is derived from the process' information.

Its methods of invalidating this are either an atfork() handler, or
when the result of getpid() changes.

Unfortunately systemd has to use a raw_clone() call, because the
high-level binding in glibc is not appropriate, so the atfork() handlers
are not called, and the cached result of getpid() is not invalidated, so
libselinux still thinks it is the old process, which means it tries to
change the exec context for the wrong process.

The long term fixes to this are to either have libselinux not cache this
information, or have glibc provide a binding to clone() that invalidates
its internal caching.