resolved: fallback to TCP if UDP fails
This is inspired by the logic in BIND [0], follow-up patches will implement the reset of that scheme.

If we get a server error back, or if after two attempts we don't get a reply at all, we switch from UDP to
TCP for the given server for the current and all subsequent requests. However, if we ever successfully
received a reply over UDP, we never fall back to TCP, and once a timeout (currently six hours) has passed,
we try to upgrade again to using UDP.

Note that UDP is likely much more commonly supported by TCP, but depending on the path between the client
and the server, we may have more luck with TCP in case something is wrong. We really do prefer UDP though
as that is much more lightweight, that is why TCP is only the last resort.

A quirk of this patch is that the server to be used is determined in the transaction, rather than in the
scope (where it arguably belongs). The reason for this is that in follow-up patches the actual DNS packet
will depend on the feature level of the server being used, so we need to determine the server before we
emmit the packet.

[0]: <https://kb.isc.org/article/AA-01219/0/Refinements-to-EDNS-fallback-behavior-can-cause-different-outcomes-in-Recursive-Servers.html>