logind: honor PowerKeyIgnoreInhibited and friends
Wording in man 5 logind.conf for respective options suggests, that
if *IgnoreInhibited=yes then requested action is executed regardless whether
inhibitor is taken or not. However now this is not the case. For example
consider inhibitor taken by gdm.

     Who: gdm (UID 42/gdm, PID 1465/gnome-settings-)
    What: handle-power-key:handle-suspend-key:handle-hibernate-key
     Why: GNOME handling keypresses
    Mode: block

It prevents users to initiate shutdown if power key is pressed, despite the fact
that PowerKeyIgnoreInhibited=yes.