resolved: transaction - move DNS UDP socket creation to the scope
With access to the server when creating the socket, we can connect()
to the server and hence simplify message sending by using write()
instead of sendmsg(). Moreover, we no longer need to verify the
address/port of received packets as these will be enforced by the
kernel.

Lastly, we no longer request PKTINFO for Unicast DNS, as this was
only ever used by LLMNR.